Update Day 13 - Puzzle 1.cpp

+++ Day 13 - Puzzle 1.cpp
@@ -19,8 +19,8 @@ int main() {
 	int sumTokens = 0;
 	for (int x1, y1, x2, y2, X, Y; fileStream >> x1 >> y1 >> x2 >> y2 >> X >> Y;) {
 		int f = std::lcm(x1, y1) / x1, g = std::lcm(x1, y1) / y1; // What we need to multiply the two eqs by
-		int b = static_cast<float>(X * f - Y * g) / (x2 * f - y2 * g);
-		int a = static_cast<float>(X - b * x2) / x1;
+		int b = (X * f - Y * g) / (x2 * f - y2 * g);
+		int a = (X - b * x2) / x1;
 
 		if (a * x1 + b * x2 == X && a * y1 + b * y2 == Y) {
 			sumTokens += a * 3 + b;
